chore: Fix typos and casing errors in codebase
This commit includes changes to fix several typos and casing errors in the codebase. The changes were made to improve the readability and consistency of the code. Specifically, the following changes were made:

- Corrected typos and replaced arbitrary CSS values with custom values defined in tailwind.config.js to ensure consistency and maintainability of the codebase. Specifically, updated values for the font-size and margin properties in the home page componentsailwind.config.js customisation
- Corrected casing errors in assets folder for the maintainbility and consitency
- Standardized naming conventions across the codebase

These changes have no impact on the functionality of the code, but they improve the overall quality of the codebase and make it easier for other developers to read and understand the code.